+++ src/convert/trits_test.go
@@ -25,7 +25,7 @@ func TestBytesToTrits(t *testing.T) {
 
 func TestTritsToInt(t *testing.T) {
 	trits := TrytesToTrits("HELLOWORLD")
-	expected := 33792688625192
+	expected := int64(33792688625192)
 	result := TritsToInt(trits)
 	if !reflect.DeepEqual(expected, result) {
 		t.Error("Trits int converion wrong!", expected, result)
